added notes and minor refactor
