+++ ProjectMgmt/src/WorkOrderMain.java
@@ -1,0 +1,8 @@
+
+public class WorkOrderMain {
+	
+	public static void main(String[] args) {
+		new WorkOrderMgmt();
+	}
+	
+}
